requirements.txt
moviepy==1.0.3
edge-tts
google-genai
requests
python-dotenv
pillow
numpy
proglog
imageio
imageio-ffmpeg
decorator
scipy
openai-whisper
yt-dlp
